email_web_scraper/requirements.txt
python-dotenv==0.14.0
email
selenium
python-dotenv